Refactor: sync a5 scheduler extraction with a2a3
Apply the same scheduler refactoring to a5 as done in a2a3:

- Create runtime/scheduler/ with SchedulerContext class and extracted
  methods (scheduler_types.h, scheduler_context.h, scheduler_completion.cpp,
  scheduler_cold_path.cpp, scheduler_dispatch.cpp)
- Move pto_scheduler.h/.cpp into runtime/scheduler/
- Move runtime.cpp and pto_shared_memory.cpp into runtime/shared/
- Update build_config.py: host source_dirs uses "runtime/shared"
- Switch CMake from GLOB to GLOB_RECURSE in all platform CMakeLists.txt
- Update include paths to "scheduler/pto_scheduler.h"
- Slim aicpu_executor.cpp from ~2854 to ~1244 lines
- Add missing copyright headers to onboard CMakeLists.txt files

a5-specific adaptations:
- RUNTIME_MAX_WORKER=108 (vs a2a3's 72) in scheduler_context.h
- CoreExecState profiling layout differs (no dispatch_count field)
- LocalContext uses s_block_idx/s_block_num (vs block_idx/block_num)
- No perf_aicpu_switch_buffer or perf_aicpu_flush_buffers calls
