+++ src/a5/runtime/tensormap_and_ringbuffer/runtime/pto_orchestrator.h
@@ -31,7 +31,7 @@
 #include "pto_ring_buffer.h"
 #include "pto_runtime2_types.h"
 #include "pto_submit_types.h"
-#include "pto_scheduler.h"
+#include "scheduler/pto_scheduler.h"
 #include "pto_shared_memory.h"
 #include "pto_tensormap.h"
 #include "pto_types.h"
